# swag/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API для сервиса mpit2026-reg.\n\nОсновные возможности:\n- авторизация через MAX mini app и локальный JWT\n- получение и обновление профиля текущего пользователя\n- создание событий, получение invite token и вход по нему\n- управление гостями и фиксация статуса посещения\n- wishlist и фото внутри события\n\nАутентификация:\n- защищённые методы требуют заголовок Authorization: Bearer \u003cjwt\u003e\n- JWT можно получить через /auth/max/login или через мобильный flow /auth/max/start -\u003e /auth/max/complete -\u003e /auth/max/exchange\n\nSwagger UI доступен по адресу /api/docs/.",
        "title": "mpit2026-reg API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "mpit-bot.kostya1024.ru",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/events/{eventID}/photos": {
            "get": {
                "description": "Возвращает список URL фотографий, привязанных к событию.\nДоступ к фотографиям есть только у пользователей, которым разрешено читать событие и связанные с ним материалы.\nЕсли у пользователя недостаточно прав, backend вернёт 403.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Получить фотографии события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Массив URL фотографий события",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав для просмотра фотографий",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка при получении списка фотографий",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{eventID}/photos/upload": {
            "post": {
                "description": "Принимает multipart/form-data с набором файлов и загружает фотографии в контекст указанного события.\nЗа один запрос можно загрузить не более 10 файлов.\nМетод предназначен для гостей и организаторов события; размер входного multipart-пакета ограничен backend-логикой.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Загрузить фотографии события",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор события",
                        "name": "eventID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Один или несколько файлов фотографий",
                        "name": "photos",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Фотографии успешно загружены",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "207": {
                        "description": "Часть файлов загрузилась, часть завершилась ошибкой",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "400": {
                        "description": "Некорректный eventID, отсутствие файлов или превышение лимита количества",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "JWT отсутствует или невалиден",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    },
                    "413": {
                        "description": "Слишком большой multipart-запрос",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}